feat: fast-execuite as plugin
