+++ src/channel/rpc_0_10_0.ts
@@ -9,7 +9,6 @@ import {
   AccountInvocations,
   BigNumberish,
   BlockIdentifier,
-  BlockTag,
   Call,
   DeclareContractTransaction,
   DeployAccountContractTransaction,
@@ -21,7 +20,6 @@ import {
   RPC_ERROR,
   RpcProviderOptions,
   waitForTransactionOptions,
-  type fastWaitForTransactionOptions,
 } from '../types';
 import assert from '../utils/assert';
 import { ETransactionType, JRPC, RPCSPEC010 as RPC } from '../types/api';
@@ -569,60 +567,6 @@ export class RpcChannel {
       await wait(retryInterval);
     }
     return txReceipt as RPC.TXN_RECEIPT;
-  }
-
-  public async fastWaitForTransaction(
-    txHash: BigNumberish,
-    address: string,
-    initNonceBN: BigNumberish,
-    options?: fastWaitForTransactionOptions
-  ): Promise<boolean> {
-    const initNonce = BigInt(initNonceBN);
-    let retries = options?.retries ?? 50;
-    const retryInterval = options?.retryInterval ?? 500; // 0.5s
-    const errorStates: string[] = [RPC.ETransactionExecutionStatus.REVERTED];
-    const successStates: string[] = [
-      RPC.ETransactionFinalityStatus.ACCEPTED_ON_L2,
-      RPC.ETransactionFinalityStatus.ACCEPTED_ON_L1,
-      RPC.ETransactionFinalityStatus.PRE_CONFIRMED,
-    ];
-    let txStatus: RPC.TransactionStatus;
-    const start = new Date().getTime();
-    while (retries > 0) {
-      // eslint-disable-next-line no-await-in-loop
-      await wait(retryInterval);
-
-      // eslint-disable-next-line no-await-in-loop
-      txStatus = await this.getTransactionStatus(txHash);
-      logger.info(
-        `${retries} ${JSON.stringify(txStatus)} ${(new Date().getTime() - start) / 1000}s.`
-      );
-      const executionStatus = txStatus.execution_status ?? '';
-      const finalityStatus = txStatus.finality_status;
-      if (errorStates.includes(executionStatus)) {
-        const message = `${executionStatus}: ${finalityStatus}`;
-        const error = new Error(message) as Error & { response: RPC.TransactionStatus };
-        error.response = txStatus;
-        throw error;
-      } else if (successStates.includes(finalityStatus)) {
-        let currentNonce = initNonce;
-        while (currentNonce === initNonce && retries > 0) {
-          // eslint-disable-next-line no-await-in-loop
-          currentNonce = BigInt(await this.getNonceForAddress(address, BlockTag.PRE_CONFIRMED));
-          logger.info(
-            `${retries} Checking new nonce ${currentNonce} ${(new Date().getTime() - start) / 1000}s.`
-          );
-          if (currentNonce !== initNonce) return true;
-          // eslint-disable-next-line no-await-in-loop
-          await wait(retryInterval);
-          retries -= 1;
-        }
-        return false;
-      }
-
-      retries -= 1;
-    }
-    return false;
   }
 
   public getStorageAt(
