+++ src/types/lib/index.ts
@@ -342,11 +342,6 @@ export type waitForTransactionOptions = {
   errorStates?: Array<TransactionFinalityStatus | TransactionExecutionStatus>;
 };
 
-export type fastWaitForTransactionOptions = {
-  retries?: number;
-  retryInterval?: number;
-};
-
 export type getSimulateTransactionOptions = {
   blockIdentifier?: BlockIdentifier;
   skipValidate?: boolean;
